LEPP-241: Update response models to make handling NPS data easier.

+++ app/models/backend/SubmitLeppRequest.scala
@@ -19,7 +19,7 @@ package models.backend
 import models.userAnswers.BankAccountDetails
 import play.api.libs.json.{Json, OWrites}
 
-case class SubmitLeppRequest(currentLowEarnersOptimisticLock: Int,
+case class SubmitLeppRequest(currentLowEarnersOptimisticLock: BigInt,
                              taxYear: Int,
                              accountDetails: BankAccountDetails)
