+++ app/models/backend/SubmitLeppResponse.scala
@@ -18,7 +18,7 @@ package models.backend
 
 import play.api.libs.json.{Json, Reads}
 
-case class SubmitLeppResponse(updatedLowEarnersOptimisticLock: Int)
+case class SubmitLeppResponse(updatedLowEarnersOptimisticLock: BigInt)
 
 object SubmitLeppResponse {
   implicit val reads: Reads[SubmitLeppResponse] = Json.reads[SubmitLeppResponse]
